# perf: 80% faster `doc.get` for fields with `None` as value