## Issue

`doc.get` wasn't performant for fields with value as `None`. Additionally, it set a value for child table fields if one didn't exist. This is unexpected behavior in a getter function.

The reason why `doc.get` did this is so that one could confidently loop over `doc.get(child_fieldname)` without worrying about the object being non-iterable.

## Solution
- Set all child fieldnames without values to an empty list whenever a Document is initialized.
- Don't do anything special in `doc.get`

## Screenshots

### Before

```python
In [5]: %timeit doc.get("from_date")
1.25 µs ± 182 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)
```

### After (~80% improvement)

```python
In [4]: %timeit doc.get("from_date")
162 ns ± 2.26 ns per loop (mean ± std. dev. of 7 runs, 10000000 loops each)

```

This might not seem like a lot, but it has significant impact on a lot of utilities like `doc.as_dict`

## Other Changes
- Update `doc.append` to only set the value of child field if it is None. Previously, it was doing so for all falsy values, but that includes an empty list (set in `doc.set` or `doc.init_child_tables`), which should not be reset.
- Assume truthy value to be iterable in `get_all_children`
- Define `DOCTYPE_TABLE_FIELDS` in `base_document.py`, import in `meta.py`
- Import and use `_dict` directly instead of `frappe._dict`
- Reformatting for readability

Note: `init_child_tables` isn't needed for Document initialized via `load_from_db` route, that happens already.